generating vertex coordinates... will this actually give a speedup?
