+++ test_autoarray/structures/triangles/coordinate/test_vertex_coordinates.py
@@ -1,0 +1,6 @@
+def test_vertex_coordinates(one_triangle):
+    print(one_triangle.vertex_coordinates)
+
+
+def test_vertex_coordinates_two_triangles(two_triangles):
+    print(two_triangles.vertex_coordinates)
